+++ crates/stackable-operator/CHANGELOG.md
@@ -8,7 +8,13 @@ All notable changes to this project will be documented in this file.
 
 - Add support for specifying a `clientAuthenticationMethod` for OIDC ([#1178]).
   This was originally done in [#1158] and had been reverted in [#1170].
+- Add generic database connection mechanism ([#1163])
 
+### Changed
+
+- BREAKING: Change signature of `ContainerBuilder::add_env_vars` from `Vec<EnvVar>` to `IntoIterator<Item = EnvVar>` ([#1163])
+
+[#1163]: https://github.com/stackabletech/operator-rs/pull/1163
 [#1178]: https://github.com/stackabletech/operator-rs/pull/1178
 
 ## [0.108.0] - 2026-03-10